Improved abstractions and structure of Tile
